requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
emoji==1.2.0
filelock==3.13.1
fsspec==2024.2.0
grpcio==1.62.1
grpcio-tools==1.62.1
huggingface-hub==0.21.4
idna==3.6
iniconfig==2.0.0
kss==4.5.4
networkx==3.2.1
numpy==1.26.4
obs-websocket-py==1.0
packaging==24.0
pecab==1.0.8
pluggy==1.4.0
protobuf==4.25.3
pyarrow==15.0.1
PyAudio==0.2.14
pybind11==2.9.2
pytest==8.1.1
python-dotenv==1.0.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
safetensors==0.4.2
tokenizers==0.15.2
tqdm==4.66.2
transformers==4.38.2
typing_extensions==4.10.0
urllib3==2.2.1
websocket-client==1.7.0
